Separate OpenJij preparation responsibilities
