+++ python/ommx-openjij-adapter/ommx_openjij_adapter/_decode.py
@@ -3,16 +3,16 @@ from __future__ import annotations
 from collections.abc import Mapping
 
 import openjij as oj
-from ommx import Samples, State
+from ommx import Instance, Samples, SampleSet, State
 
 
-def _decode_to_samples(
+def _decode_response(
     response: oj.Response,
     *,
     variable_ids: set[int] | None = None,
     default_values: Mapping[int, float] | None = None,
 ) -> Samples:
-    """Convert an OpenJij response without evaluating an OMMX Instance."""
+    """Decode OpenJij response records into OMMX samples."""
     samples = Samples({})
     sample_id = 0
     filtered_defaults = {
@@ -42,8 +42,19 @@ def _decode_to_samples(
     return samples
 
 
+def _decode_for_instance(response: oj.Response, instance: Instance) -> SampleSet:
+    """Decode and evaluate a response against the exact Adapter input."""
+    variable_ids = {variable.id for variable in instance.used_decision_variables}
+    samples = _decode_response(
+        response,
+        variable_ids=variable_ids,
+        default_values={variable_id: 0.0 for variable_id in variable_ids},
+    )
+    return instance.evaluate_samples(samples)
+
+
 def decode_to_samples(response: oj.Response) -> Samples:
     """
     Convert `openjij.Response <https://openjij.github.io/OpenJij/reference/openjij/index.html#openjij.Response>`_ to :class:`Samples`
     """
-    return _decode_to_samples(response)
+    return _decode_response(response)
