+++ python/ommx-openjij-adapter/ommx_openjij_adapter/_preparation.py
@@ -1,35 +1,18 @@
-"""OpenJij-specific preparation planning and materialization."""
+"""OpenJij-specific preparation results and diagnostics."""
 
 from __future__ import annotations
 
-from ommx import (
-    AdditionalCapability,
-    DegreeBound,
-    Equality,
-    Instance,
-    InstanceClass,
-    InstanceClassClause,
-    InstanceClassMembershipReport,
-    Kind,
-    Sense,
-    State,
-    Samples,
-    SampleSet,
-)
+import copy
+from dataclasses import dataclass, field
+
+from ommx import Instance, InstanceClassMembershipReport, Samples, SampleSet, State
 from ommx.adapter import (
     AdapterApplicabilityReport,
     AdapterPreconditionViolation,
     ConstraintRef,
-    InfeasibleDetected,
 )
-from opentelemetry import trace
-from collections.abc import Callable, Iterable, Mapping
-from dataclasses import dataclass, field
-from math import isfinite
-from typing import ClassVar
-import copy
 
-_tracer = trace.get_tracer("ommx.adapter.openjij")
+PreparationDiagnosticValue = str | int | float | bool | None
 
 
 @dataclass(frozen=True, slots=True)
@@ -74,17 +57,31 @@ class OpenJijPreparationSourceCheck:
 
 
 @dataclass(frozen=True, slots=True)
+class OpenJijPreparationFailure:
+    """One failure discovered while materializing an accepted source."""
+
+    reason: str
+    description: str
+    variable_ids: frozenset[int] = field(default_factory=frozenset)
+    constraint_refs: frozenset[ConstraintRef] = field(default_factory=frozenset)
+    observed: PreparationDiagnosticValue = None
+    expected: PreparationDiagnosticValue = None
+
+
+@dataclass(frozen=True, slots=True)
 class OpenJijPreparationReport:
-    """OpenJij operation audit and applicability of the produced Adapter input."""
+    """Checks, operation audit, failures, and produced-input applicability."""
 
     source_check: OpenJijPreparationSourceCheck
     steps: tuple[OpenJijPreparationStep, ...]
-    input_applicability: AdapterApplicabilityReport | None
+    preparation_failures: tuple[OpenJijPreparationFailure, ...] = ()
+    input_applicability: AdapterApplicabilityReport | None = None
 
     @property
     def is_successful(self) -> bool:
         return (
             self.source_check.conditions_hold
+            and not self.preparation_failures
             and self.input_applicability is not None
             and self.input_applicability.is_applicable
         )
@@ -97,6 +94,10 @@ class OpenJijPreparation:
     _input: Instance = field(repr=False)
     _source_instance: Instance = field(repr=False)
     report: OpenJijPreparationReport
+
+    def __post_init__(self) -> None:
+        if not self.report.is_successful:
+            raise ValueError("OpenJijPreparation requires a successful report")
 
     @property
     def input(self) -> Instance:
@@ -147,808 +148,12 @@ class OpenJijPreparationError(ValueError):
                 for violation in source_check.precondition_violations
             )
             message = f"OpenJij preparation preconditions failed:\n{details}"
+        elif report.preparation_failures:
+            details = "\n".join(
+                f"- {failure.reason}: {failure.description}"
+                for failure in report.preparation_failures
+            )
+            message = f"OpenJij preparation failed:\n{details}"
         else:
             message = "OpenJij preparation did not produce an applicable input"
         super().__init__(message)
-
-
-class _OpenJijPreparationSupport:
-    """Internal support for explicit OpenJij input preparation."""
-
-    @classmethod
-    def _check_prepared_input_applicability(
-        cls, ommx_instance: Instance
-    ) -> AdapterApplicabilityReport:
-        """Adapter-owned callback for validating the produced input."""
-        raise NotImplementedError
-
-    # This describes sources accepted by the explicit preparation operation,
-    # not inputs accepted directly by the OpenJij Adapter.
-    _PREPARATION_SOURCE_CLASS: ClassVar[InstanceClass] = InstanceClass(
-        [
-            InstanceClassClause(
-                label="openjij-preparation-source",
-                allowed_variable_kinds={Kind.Binary, Kind.Integer},
-                objective_degree_bound=DegreeBound.unbounded(),
-                regular_constraint_degree_bounds={
-                    Equality.EqualToZero: DegreeBound.unbounded(),
-                    Equality.LessThanOrEqualToZero: DegreeBound.unbounded(),
-                },
-                indicator_constraint_degree_bounds={
-                    Equality.EqualToZero: DegreeBound.unbounded(),
-                    Equality.LessThanOrEqualToZero: DegreeBound.unbounded(),
-                },
-                allows_one_hot=True,
-                allows_sos1=True,
-                allowed_senses={Sense.Minimize, Sense.Maximize},
-            )
-        ]
-    )
-    _ENCODING_INPUT_CLASS: ClassVar[InstanceClass] = InstanceClass(
-        [
-            InstanceClassClause(
-                label="openjij-log-encoding-input",
-                allowed_variable_kinds={Kind.Binary, Kind.Integer},
-                objective_degree_bound=DegreeBound.unbounded(),
-                allowed_senses={Sense.Minimize},
-            )
-        ]
-    )
-    MAX_LOG_ENCODING_BITS: ClassVar[int] = 53
-    MAX_SLACK_RANGE: ClassVar[int] = 2**64 - 1
-
-    @staticmethod
-    def _active_constraint_refs(ommx_instance: Instance) -> frozenset[ConstraintRef]:
-        return frozenset(
-            [
-                *(ConstraintRef("regular", id) for id in ommx_instance.constraints),
-                *(
-                    ConstraintRef("indicator", id)
-                    for id in ommx_instance.indicator_constraints
-                ),
-                *(
-                    ConstraintRef("one_hot", id)
-                    for id in ommx_instance.one_hot_constraints
-                ),
-                *(ConstraintRef("sos1", id) for id in ommx_instance.sos1_constraints),
-            ]
-        )
-
-    @classmethod
-    def _check_class_preconditions(
-        cls,
-        ommx_instance: Instance,
-        input_class: InstanceClass,
-        check_preconditions: Callable[[], Iterable[AdapterPreconditionViolation]],
-    ) -> tuple[
-        InstanceClassMembershipReport,
-        bool,
-        tuple[AdapterPreconditionViolation, ...],
-    ]:
-        membership = input_class.check_membership(ommx_instance)
-        if not membership.is_member:
-            return membership, False, ()
-        return membership, True, tuple(check_preconditions())
-
-    @classmethod
-    def _log_encoding_precondition_violations(
-        cls, ommx_instance: Instance
-    ) -> tuple[AdapterPreconditionViolation, ...]:
-        encoding_input = copy.deepcopy(ommx_instance)
-        try:
-            encoding_input.reduce_capabilities(set())
-        except Exception:
-            # Full preparation materialization reports special-constraint
-            # lowering failures with their source constraint references.
-            return ()
-
-        integer_variables = {
-            variable.id: variable
-            for variable in encoding_input.used_decision_variables
-            if variable.kind == Kind.Integer
-        }
-        if not integer_variables:
-            return ()
-
-        violations: list[AdapterPreconditionViolation] = []
-        max_exact_integer = float(2**cls.MAX_LOG_ENCODING_BITS)
-        max_range_width = float(2**cls.MAX_LOG_ENCODING_BITS - 1)
-        for variable_id, variable in integer_variables.items():
-            candidate = copy.deepcopy(encoding_input)
-            try:
-                candidate.log_encode({variable_id})
-            except Exception as error:
-                lower = variable.bound.lower
-                upper = variable.bound.upper
-                width = upper - lower
-                description = (
-                    f"Integer variable {variable_id} cannot be log-encoded for "
-                    f"OpenJij preparation: {error}"
-                )
-                if not isfinite(lower) or not isfinite(upper):
-                    condition = "openjij.log_encoding.bound_finite"
-                    actual: str | int | float = f"[{lower}, {upper}]"
-                    limit: str | int | float = "finite integer range"
-                elif width > max_range_width:
-                    condition = "openjij.log_encoding.max_bits"
-                    actual = int(width).bit_length()
-                    limit = cls.MAX_LOG_ENCODING_BITS
-                elif width != 0 and (
-                    lower < -max_exact_integer or upper > max_exact_integer
-                ):
-                    condition = "openjij.log_encoding.exact_integer_range"
-                    actual = lower if lower < -max_exact_integer else upper
-                    limit = max_exact_integer
-                else:
-                    condition = "openjij.log_encoding.failed"
-                    actual = f"[{lower}, {upper}]"
-                    limit = (
-                        "finite unit-spaced range encodable with at most "
-                        f"{cls.MAX_LOG_ENCODING_BITS} bits"
-                    )
-                violations.append(
-                    AdapterPreconditionViolation(
-                        condition=condition,
-                        description=description,
-                        variable_ids=frozenset({variable_id}),
-                        actual=actual,
-                        limit=limit,
-                    )
-                )
-
-        if violations:
-            return tuple(violations)
-
-        candidate = copy.deepcopy(encoding_input)
-        try:
-            candidate.log_encode(set(integer_variables))
-        except Exception as error:
-            return (
-                AdapterPreconditionViolation(
-                    condition="openjij.log_encoding.combined_rewrite",
-                    description=(
-                        "The complete Integer-to-Binary rewrite cannot be applied "
-                        f"for OpenJij preparation: {error}"
-                    ),
-                    variable_ids=frozenset(integer_variables),
-                    actual=str(error),
-                    limit="a finite atomic log-encoding rewrite",
-                ),
-            )
-        return ()
-
-    @classmethod
-    def _penalty_precondition_violations(
-        cls,
-        ommx_instance: Instance,
-        uniform_penalty_weight: float | None,
-        penalty_weights: Mapping[int, float] | None,
-        inequality_integer_slack_max_range: int,
-    ) -> tuple[AdapterPreconditionViolation, ...]:
-        constraint_ids = frozenset(ommx_instance.constraints)
-        constraint_refs = cls._active_constraint_refs(ommx_instance)
-        special_constraint_refs = frozenset(
-            ref for ref in constraint_refs if ref.family != "regular"
-        )
-        violations: list[AdapterPreconditionViolation] = []
-
-        if uniform_penalty_weight is not None and penalty_weights is not None:
-            violations.append(
-                AdapterPreconditionViolation(
-                    condition="openjij.penalty.options_exclusive",
-                    description=(
-                        "Choose either a uniform penalty weight or per-constraint "
-                        "penalty weights, not both."
-                    ),
-                    constraint_refs=constraint_refs,
-                    actual="both selected",
-                    limit="one penalty mode",
-                )
-            )
-
-        if not constraint_refs and (
-            uniform_penalty_weight is not None or penalty_weights is not None
-        ):
-            violations.append(
-                AdapterPreconditionViolation(
-                    condition="openjij.penalty.unused",
-                    description="Penalty weights were supplied for an unconstrained model.",
-                    actual="penalty weights supplied",
-                    limit="no penalty configuration",
-                )
-            )
-
-        if uniform_penalty_weight is not None and (
-            not isfinite(uniform_penalty_weight) or uniform_penalty_weight <= 0
-        ):
-            violations.append(
-                AdapterPreconditionViolation(
-                    condition="openjij.penalty.weight_positive_finite",
-                    description="The uniform penalty weight must be finite and positive.",
-                    constraint_refs=constraint_refs,
-                    actual=uniform_penalty_weight,
-                    limit="finite value > 0",
-                )
-            )
-
-        if penalty_weights is not None:
-            if special_constraint_refs:
-                violations.append(
-                    AdapterPreconditionViolation(
-                        condition="openjij.penalty.special_requires_uniform",
-                        description=(
-                            "Per-constraint weights are keyed by regular constraint ID "
-                            "and cannot identify constraints introduced by exact special-"
-                            "constraint lowering. Use a uniform penalty weight."
-                        ),
-                        constraint_refs=special_constraint_refs,
-                        actual="per-constraint penalty weights",
-                        limit="uniform_penalty_weight",
-                    )
-                )
-            missing = constraint_ids.difference(penalty_weights)
-            if missing:
-                violations.append(
-                    AdapterPreconditionViolation(
-                        condition="openjij.penalty.weight_coverage",
-                        description=(
-                            "Per-constraint penalty weights do not cover every "
-                            f"active regular constraint: missing {sorted(missing)}."
-                        ),
-                        constraint_refs=frozenset(
-                            ConstraintRef("regular", constraint_id)
-                            for constraint_id in missing
-                        ),
-                        actual=len(penalty_weights),
-                        limit=len(constraint_ids),
-                    )
-                )
-            unexpected = frozenset(penalty_weights).difference(constraint_ids)
-            if unexpected:
-                violations.append(
-                    AdapterPreconditionViolation(
-                        condition="openjij.penalty.weight_coverage",
-                        description=(
-                            "Per-constraint penalty weights contain unknown regular "
-                            f"constraint IDs: {sorted(unexpected)}."
-                        ),
-                        constraint_refs=frozenset(
-                            ConstraintRef("regular", constraint_id)
-                            for constraint_id in unexpected
-                        ),
-                        actual=len(penalty_weights),
-                        limit=len(constraint_ids),
-                    )
-                )
-            for constraint_id, weight in penalty_weights.items():
-                if not isfinite(weight) or weight <= 0:
-                    violations.append(
-                        AdapterPreconditionViolation(
-                            condition="openjij.penalty.weight_positive_finite",
-                            description=(
-                                f"Penalty weight for regular constraint {constraint_id} "
-                                "must be finite and positive."
-                            ),
-                            constraint_refs=frozenset(
-                                {ConstraintRef("regular", constraint_id)}
-                            ),
-                            actual=weight,
-                            limit="finite value > 0",
-                        )
-                    )
-
-        inequality_refs = frozenset(
-            ConstraintRef("regular", constraint_id)
-            for constraint_id, constraint in ommx_instance.constraints.items()
-            if constraint.equality == Equality.LessThanOrEqualToZero
-        )
-        slack_relevant_refs = inequality_refs.union(
-            ref
-            for ref in special_constraint_refs
-            if ref.family in {"indicator", "sos1"}
-        )
-        valid_slack_range = (
-            isinstance(inequality_integer_slack_max_range, int)
-            and not isinstance(inequality_integer_slack_max_range, bool)
-            and 0 < inequality_integer_slack_max_range <= cls.MAX_SLACK_RANGE
-        )
-        if slack_relevant_refs and not valid_slack_range:
-            violations.append(
-                AdapterPreconditionViolation(
-                    condition="openjij.slack.range_unsigned_64_bit",
-                    description=(
-                        "The integer slack range must fit the positive unsigned "
-                        "64-bit range when preparing inequality constraints."
-                    ),
-                    constraint_refs=slack_relevant_refs,
-                    actual=inequality_integer_slack_max_range,
-                    limit=f"integer in [1, {cls.MAX_SLACK_RANGE}]",
-                )
-            )
-        return tuple(violations)
-
-    @classmethod
-    def _check_preparation_source(
-        cls,
-        ommx_instance: Instance,
-        *,
-        uniform_penalty_weight: float | None = None,
-        penalty_weights: Mapping[int, float] | None = None,
-        inequality_integer_slack_max_range: int = 32,
-    ) -> OpenJijPreparationSourceCheck:
-        membership, preconditions_checked, violations = cls._check_class_preconditions(
-            ommx_instance,
-            cls._PREPARATION_SOURCE_CLASS,
-            lambda: (
-                *cls._log_encoding_precondition_violations(ommx_instance),
-                *cls._penalty_precondition_violations(
-                    ommx_instance,
-                    uniform_penalty_weight,
-                    penalty_weights,
-                    inequality_integer_slack_max_range,
-                ),
-            ),
-        )
-        return OpenJijPreparationSourceCheck(
-            source_membership=membership,
-            preconditions_checked=preconditions_checked,
-            precondition_violations=violations,
-        )
-
-    @classmethod
-    def _plan_preparation(
-        cls,
-        ommx_instance: Instance,
-        *,
-        uniform_penalty_weight: float | None,
-        penalty_weights: Mapping[int, float] | None,
-        inequality_integer_slack_max_range: int,
-        allow_approximate_integer_slack: bool,
-    ) -> tuple[OpenJijPreparationReport, OpenJijPreparation | None]:
-        source_check = cls._check_preparation_source(
-            ommx_instance,
-            uniform_penalty_weight=uniform_penalty_weight,
-            penalty_weights=penalty_weights,
-            inequality_integer_slack_max_range=inequality_integer_slack_max_range,
-        )
-        if (
-            not source_check.source_membership.is_member
-            or source_check.precondition_violations
-        ):
-            return (
-                OpenJijPreparationReport(
-                    source_check=source_check,
-                    steps=(),
-                    input_applicability=None,
-                ),
-                None,
-            )
-
-        try:
-            prepared = cls._materialize_preparation(
-                ommx_instance,
-                source_check=source_check,
-                uniform_penalty_weight=uniform_penalty_weight,
-                penalty_weights=penalty_weights,
-                inequality_integer_slack_max_range=inequality_integer_slack_max_range,
-                allow_approximate_integer_slack=allow_approximate_integer_slack,
-            )
-        except InfeasibleDetected:
-            raise
-        except OpenJijPreparationError as error:
-            return error.report, None
-        except Exception as error:
-            violation = AdapterPreconditionViolation(
-                condition="openjij.preparation.materialization",
-                description=(
-                    "The explicit OpenJij preparation transformations could not "
-                    f"be materialized on an isolated copy: {error}"
-                ),
-                variable_ids=frozenset(
-                    variable.id for variable in ommx_instance.used_decision_variables
-                ),
-                constraint_refs=cls._active_constraint_refs(ommx_instance),
-                actual=str(error),
-                limit="a successfully materialized prepared input",
-            )
-            return (
-                OpenJijPreparationReport(
-                    source_check=OpenJijPreparationSourceCheck(
-                        source_membership=source_check.source_membership,
-                        preconditions_checked=source_check.preconditions_checked,
-                        precondition_violations=(
-                            *source_check.precondition_violations,
-                            violation,
-                        ),
-                    ),
-                    steps=(),
-                    input_applicability=None,
-                ),
-                None,
-            )
-        return prepared.report, prepared
-
-    @classmethod
-    def check_preparation(
-        cls,
-        ommx_instance: Instance,
-        *,
-        uniform_penalty_weight: float | None = None,
-        penalty_weights: Mapping[int, float] | None = None,
-        inequality_integer_slack_max_range: int = 32,
-        allow_approximate_integer_slack: bool = False,
-    ) -> OpenJijPreparationReport:
-        """Dry-run the complete explicit preparation without mutating the input.
-
-        This is intentionally separate from :meth:`check_applicability`, which
-        checks only the Binary, unconstrained minimization Adapter
-        input. The 53-bit log-encoding limit is a preparation precondition, not
-        an OpenJij input-class condition and not an ``ommx.v2.Feature``. A model
-        proven infeasible while planning integer slack raises
-        :class:`~ommx.adapter.InfeasibleDetected`. Approximate integer slack is
-        disabled unless ``allow_approximate_integer_slack=True`` is supplied.
-        """
-
-        report, _ = cls._plan_preparation(
-            ommx_instance,
-            uniform_penalty_weight=uniform_penalty_weight,
-            penalty_weights=penalty_weights,
-            inequality_integer_slack_max_range=inequality_integer_slack_max_range,
-            allow_approximate_integer_slack=allow_approximate_integer_slack,
-        )
-        return report
-
-    @classmethod
-    def _check_encoding_input(
-        cls, ommx_instance: Instance
-    ) -> tuple[
-        InstanceClassMembershipReport,
-        bool,
-        tuple[AdapterPreconditionViolation, ...],
-    ]:
-        return cls._check_class_preconditions(
-            ommx_instance,
-            cls._ENCODING_INPUT_CLASS,
-            lambda: cls._log_encoding_precondition_violations(ommx_instance),
-        )
-
-    @classmethod
-    def prepare(
-        cls,
-        ommx_instance: Instance,
-        *,
-        uniform_penalty_weight: float | None = None,
-        penalty_weights: Mapping[int, float] | None = None,
-        inequality_integer_slack_max_range: int = 32,
-        allow_approximate_integer_slack: bool = False,
-    ) -> OpenJijPreparation:
-        """Produce a separate Adapter input and an auditable preparation report.
-
-        Raises :class:`~ommx.adapter.InfeasibleDetected` when variable bounds
-        prove an inequality infeasible. Other preparation failures
-        raise :class:`OpenJijPreparationError`. Approximate integer slack is
-        used only when ``allow_approximate_integer_slack=True`` is supplied.
-        """
-
-        with _tracer.start_as_current_span("prepare") as span:
-            span.set_attribute("adapter", f"{cls.__module__}.{cls.__qualname__}")
-            report, prepared = cls._plan_preparation(
-                ommx_instance,
-                uniform_penalty_weight=uniform_penalty_weight,
-                penalty_weights=penalty_weights,
-                inequality_integer_slack_max_range=inequality_integer_slack_max_range,
-                allow_approximate_integer_slack=allow_approximate_integer_slack,
-            )
-            if prepared is None:
-                raise OpenJijPreparationError(report)
-            return prepared
-
-    @classmethod
-    def _materialize_preparation(
-        cls,
-        ommx_instance: Instance,
-        *,
-        source_check: OpenJijPreparationSourceCheck,
-        uniform_penalty_weight: float | None,
-        penalty_weights: Mapping[int, float] | None,
-        inequality_integer_slack_max_range: int,
-        allow_approximate_integer_slack: bool,
-    ) -> OpenJijPreparation:
-        """Explicitly prepare a source Instance and return its separate input.
-
-        Sense reversal and Integer log encoding use exact OMMX operations.
-        Approximate integer slack requires explicit selection, and every
-        constraint penalty is finite rather than an assertion of direct or
-        exact constrained support.
-        """
-
-        source_instance = copy.deepcopy(ommx_instance)
-        working = copy.deepcopy(ommx_instance)
-        steps: list[OpenJijPreparationStep] = []
-
-        special_refs = {
-            AdditionalCapability.Indicator: frozenset(
-                ConstraintRef("indicator", id) for id in working.indicator_constraints
-            ),
-            AdditionalCapability.OneHot: frozenset(
-                ConstraintRef("one_hot", id) for id in working.one_hot_constraints
-            ),
-            AdditionalCapability.Sos1: frozenset(
-                ConstraintRef("sos1", id) for id in working.sos1_constraints
-            ),
-        }
-        converted_specials = working.reduce_capabilities(set())
-        special_step_details = {
-            AdditionalCapability.Indicator: (
-                "indicator_lowering",
-                "Lowered Indicator constraints exactly with validated Big-M bounds.",
-            ),
-            AdditionalCapability.OneHot: (
-                "one_hot_lowering",
-                "Lowered OneHot constraints exactly to regular equalities.",
-            ),
-            AdditionalCapability.Sos1: (
-                "sos1_lowering",
-                "Lowered SOS1 constraints exactly with validated Big-M bounds.",
-            ),
-        }
-        for capability in (
-            AdditionalCapability.Indicator,
-            AdditionalCapability.OneHot,
-            AdditionalCapability.Sos1,
-        ):
-            if capability not in converted_specials:
-                continue
-            operation, description = special_step_details[capability]
-            steps.append(
-                OpenJijPreparationStep(
-                    operation=operation,
-                    description=description,
-                    constraint_refs=special_refs[capability],
-                )
-            )
-
-        source_integer_ids = frozenset(
-            variable.id
-            for variable in source_instance.used_decision_variables
-            if variable.kind == Kind.Integer
-        )
-        if source_integer_ids:
-            working.log_encode(set(source_integer_ids))
-            steps.append(
-                OpenJijPreparationStep(
-                    operation="integer_log_encoding",
-                    description=(
-                        "Log-encoded source Integer variables after validating "
-                        f"the {cls.MAX_LOG_ENCODING_BITS}-bit preparation limit."
-                    ),
-                    variable_ids=source_integer_ids,
-                )
-            )
-
-        sense_reversed = working.as_minimization_problem()
-        if sense_reversed:
-            steps.append(
-                OpenJijPreparationStep(
-                    operation="sense_reversal",
-                    description=(
-                        "Negated the objective for the Adapter minimization input; sample "
-                        "evaluation retains the source maximization sense."
-                    ),
-                )
-            )
-
-        inequality_ids = [
-            constraint_id
-            for constraint_id, constraint in working.constraints.items()
-            if constraint.equality == Equality.LessThanOrEqualToZero
-        ]
-        for constraint_id in inequality_ids:
-            constraint_ref = frozenset({ConstraintRef("regular", constraint_id)})
-            try:
-                working.convert_inequality_to_equality_with_integer_slack(
-                    constraint_id, inequality_integer_slack_max_range
-                )
-            except RuntimeError as exact_error:
-                exact_message = str(exact_error)
-                if (
-                    "The bound of `f(x)` in inequality constraint" in exact_message
-                    and "is positive" in exact_message
-                ):
-                    raise InfeasibleDetected(exact_message) from None
-                if not allow_approximate_integer_slack:
-                    violation = AdapterPreconditionViolation(
-                        condition="openjij.slack.approximation_explicit_selection",
-                        description=(
-                            "Exact integer slack was unavailable "
-                            f"({exact_error}). Set "
-                            "allow_approximate_integer_slack=True to permit "
-                            "discrete slack approximation."
-                        ),
-                        constraint_refs=constraint_ref,
-                        actual="not selected",
-                        limit="allow_approximate_integer_slack=True",
-                    )
-                    failed_report = OpenJijPreparationReport(
-                        source_check=OpenJijPreparationSourceCheck(
-                            source_membership=source_check.source_membership,
-                            preconditions_checked=source_check.preconditions_checked,
-                            precondition_violations=(
-                                *source_check.precondition_violations,
-                                violation,
-                            ),
-                        ),
-                        steps=tuple(steps),
-                        input_applicability=None,
-                    )
-                    raise OpenJijPreparationError(failed_report)
-                try:
-                    residual_step = working.add_integer_slack_to_inequality(
-                        constraint_id, inequality_integer_slack_max_range
-                    )
-                except RuntimeError as approximate_error:
-                    message = str(approximate_error)
-                    if (
-                        "The bound of `f(x)` in inequality constraint" in message
-                        and "is positive" in message
-                    ):
-                        raise InfeasibleDetected(message) from None
-                    raise
-                steps.append(
-                    OpenJijPreparationStep(
-                        operation="approximate_integer_slack",
-                        description=(
-                            "Exact integer slack was unavailable "
-                            f"({exact_error}); used a discrete slack with residual "
-                            f"step {residual_step}."
-                        ),
-                        constraint_refs=constraint_ref,
-                    )
-                )
-            else:
-                if constraint_id in working.constraints:
-                    operation = "exact_integer_slack"
-                    description = "Converted the inequality with exact integer slack."
-                else:
-                    operation = "trivial_inequality_removal"
-                    description = (
-                        "Removed an inequality proven satisfied by the variable bounds."
-                    )
-                steps.append(
-                    OpenJijPreparationStep(
-                        operation=operation,
-                        description=description,
-                        constraint_refs=constraint_ref,
-                    )
-                )
-
-        remaining_constraint_ids = frozenset(working.constraints)
-        penalty_constraint_refs = frozenset(
-            [
-                *(
-                    ConstraintRef("regular", constraint_id)
-                    for constraint_id in remaining_constraint_ids
-                    if constraint_id in ommx_instance.constraints
-                ),
-                *(
-                    ref
-                    for ref in cls._active_constraint_refs(ommx_instance)
-                    if ref.family != "regular"
-                ),
-            ]
-        )
-        if remaining_constraint_ids:
-            if uniform_penalty_weight is None and penalty_weights is None:
-                violation = AdapterPreconditionViolation(
-                    condition="openjij.penalty.explicit_selection",
-                    description=(
-                        "Constraints remaining after exact preparation require an "
-                        "explicitly selected finite penalty; constrained models are "
-                        "not part of the OpenJij input class."
-                    ),
-                    constraint_refs=penalty_constraint_refs,
-                    actual="not selected",
-                    limit="uniform_penalty_weight or penalty_weights",
-                )
-                failed_report = OpenJijPreparationReport(
-                    source_check=OpenJijPreparationSourceCheck(
-                        source_membership=source_check.source_membership,
-                        preconditions_checked=source_check.preconditions_checked,
-                        precondition_violations=(
-                            *source_check.precondition_violations,
-                            violation,
-                        ),
-                    ),
-                    steps=tuple(steps),
-                    input_applicability=None,
-                )
-                raise OpenJijPreparationError(failed_report)
-            if penalty_weights is not None:
-                parametric = working.penalty_method()
-                weights: dict[int, float] = {}
-                for constraint_id in remaining_constraint_ids:
-                    removed = parametric.removed_constraints[constraint_id]
-                    parameter_id = int(
-                        removed.removed_reason_parameters["parameter_id"]
-                    )
-                    weights[parameter_id] = penalty_weights[constraint_id]
-                working = parametric.with_parameters(weights)
-                penalty_description = (
-                    "Applied positive per-constraint finite penalties."
-                )
-            else:
-                assert uniform_penalty_weight is not None
-                parametric = working.uniform_penalty_method()
-                parameter = parametric.parameters[0]
-                working = parametric.with_parameters(
-                    {parameter.id: uniform_penalty_weight}
-                )
-                penalty_description = (
-                    f"Applied finite uniform penalty weight {uniform_penalty_weight}."
-                )
-            steps.append(
-                OpenJijPreparationStep(
-                    operation="finite_penalty",
-                    description=penalty_description,
-                    constraint_refs=penalty_constraint_refs,
-                )
-            )
-
-        encoding_membership, encoding_preconditions_checked, encoding_violations = (
-            cls._check_encoding_input(working)
-        )
-        if not encoding_membership.is_member:
-            encoding_violations = (
-                AdapterPreconditionViolation(
-                    condition="openjij.preparation.encoding_input_class",
-                    description=(
-                        "The prepared intermediate value is outside the class "
-                        "supported by Integer log encoding."
-                    ),
-                    actual=str(encoding_membership),
-                    limit="Binary or Integer unconstrained minimization input",
-                ),
-            )
-        if not encoding_preconditions_checked or encoding_violations:
-            failed_report = OpenJijPreparationReport(
-                source_check=OpenJijPreparationSourceCheck(
-                    source_membership=source_check.source_membership,
-                    preconditions_checked=source_check.preconditions_checked,
-                    precondition_violations=(
-                        *source_check.precondition_violations,
-                        *encoding_violations,
-                    ),
-                ),
-                steps=tuple(steps),
-                input_applicability=None,
-            )
-            raise OpenJijPreparationError(failed_report)
-
-        slack_integer_ids = frozenset(
-            variable.id
-            for variable in working.used_decision_variables
-            if variable.kind == Kind.Integer
-        )
-        if slack_integer_ids:
-            working.log_encode(set(slack_integer_ids))
-            steps.append(
-                OpenJijPreparationStep(
-                    operation="integer_slack_log_encoding",
-                    description="Log-encoded Integer variables introduced by slack preparation.",
-                    variable_ids=slack_integer_ids,
-                )
-            )
-
-        input_applicability = cls._check_prepared_input_applicability(working)
-        report = OpenJijPreparationReport(
-            source_check=source_check,
-            steps=tuple(steps),
-            input_applicability=input_applicability,
-        )
-        if not report.is_successful:
-            raise OpenJijPreparationError(report)
-        return OpenJijPreparation(
-            _input=working,
-            _source_instance=source_instance,
-            report=report,
-        )
